code clean-up
git-svn-id: svn://svn.code.sf.net/p/mobac/code/trunk/MOBAC@2692 a036c9c4-4c23-4c51-82c3-8fdb5c987102

+++ mobac/src/main/java/mobac/program/atlascreators/GlopusMapFile.java
@@ -173,7 +173,7 @@ public class GlopusMapFile extends TrekBuddy {
                 );
                 try (LittleEndianOutputStream out = new LittleEndianOutputStream(
                         new BufferedOutputStream(fout, 16384))) {
-                    out.writeInt((int) 0xff000002);
+                    out.writeInt(0xff000002);
                     out.writeInt(count);
                     int mapNumber = 0;
                     for (GlopusTile gt : tiles) {
